+++ components/archive/impl/src/main/java/com/flipperdevices/archive/impl/composable/page/ArchiveProgressScreen.kt
@@ -11,15 +11,13 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.colorResource
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import com.flipperdevices.archive.impl.R
 import com.flipperdevices.bridge.synchronization.api.SynchronizationState
 import com.flipperdevices.core.ktx.jre.roundPercentToString
-import com.flipperdevices.core.ui.res.R as DesignSystem
+import com.flipperdevices.core.ui.theme.LocalPallet
+import com.flipperdevices.core.ui.theme.LocalTypography
 
 @Composable
 fun ArchiveProgressScreen(inProgressState: SynchronizationState.InProgress) {
@@ -30,7 +28,7 @@ fun ArchiveProgressScreen(inProgressState: SynchronizationState.InProgress) {
     ) {
         CircularProgressIndicator(
             modifier = Modifier.size(24.dp),
-            color = colorResource(DesignSystem.color.accent_secondary),
+            color = LocalPallet.current.accentSecond,
             strokeWidth = 2.dp
         )
         Text(
@@ -39,9 +37,8 @@ fun ArchiveProgressScreen(inProgressState: SynchronizationState.InProgress) {
                 R.string.archive_sync_percent,
                 inProgressState.progress.roundPercentToString()
             ),
-            fontWeight = FontWeight.W500,
-            fontSize = 14.sp,
-            color = colorResource(DesignSystem.color.black_60),
+            style = LocalTypography.current.bodyM14,
+            color = LocalPallet.current.text60,
             textAlign = TextAlign.Center
         )
     }