+++ components/archive/shared/src/main/java/com/flipperdevices/archive/shared/composable/ComposableKeyCard.kt
@@ -1,15 +1,12 @@
 package com.flipperdevices.archive.shared.composable
 
-import androidx.annotation.ColorRes
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.ColumnScope
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.selection.SelectionContainer
 import androidx.compose.material.Card
 import androidx.compose.material.Text
@@ -18,22 +15,22 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.colorResource
-import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import com.flipperdevices.archive.shared.utils.ExtractKeyMetaInformation
+import com.flipperdevices.bridge.dao.api.model.FlipperFileType.Companion.colorByFlipperFileType
 import com.flipperdevices.bridge.dao.api.model.parsed.FlipperKeyParsed
 import com.flipperdevices.core.ui.ktx.ComposableKeyType
-import com.flipperdevices.core.ui.res.R as DesignSystem
+import com.flipperdevices.core.ui.theme.LocalPallet
+import com.flipperdevices.core.ui.theme.LocalTypography
 
 @Composable
 fun ComposableKeyCard(
     modifier: Modifier,
     synchronizationContent: (@Composable () -> Unit)?,
     flipperKeyParsed: FlipperKeyParsed,
-    @ColorRes typeColorId: Int? = flipperKeyParsed.fileType?.color,
+    typeColor: Color = colorByFlipperFileType(flipperKeyParsed.fileType),
     onCardClicked: () -> Unit
 ) {
     Card(
@@ -44,13 +41,12 @@ fun ComposableKeyCard(
                 interactionSource = remember { MutableInteractionSource() },
                 indication = rememberRipple(),
                 onClick = onCardClicked
-            ),
-        shape = RoundedCornerShape(size = 10.dp)
+            )
     ) {
         Column(Modifier.padding(bottom = 8.dp)) {
             ComposableKeyCardContent(
                 flipperKeyParsed,
-                typeColorId,
+                typeColor,
                 synchronizationContent
             )
         }
@@ -58,21 +54,20 @@ fun ComposableKeyCard(
 }
 
 @Composable
-private fun ColumnScope.ComposableKeyCardContent(
+private fun ComposableKeyCardContent(
     flipperKeyParsed: FlipperKeyParsed,
-    @ColorRes typeColorId: Int?,
+    typeColor: Color,
     synchronizationContent: (@Composable () -> Unit)?
 ) {
     Row(verticalAlignment = Alignment.CenterVertically) {
-        ComposableKeyType(flipperKeyParsed.fileType, colorId = typeColorId)
+        ComposableKeyType(flipperKeyParsed.fileType, colorKey = typeColor)
         val protocol = ExtractKeyMetaInformation.extractProtocol(flipperKeyParsed)
         if (protocol != null) {
             Text(
                 modifier = Modifier.padding(horizontal = 14.dp),
                 text = protocol,
-                color = colorResource(DesignSystem.color.black_12),
-                fontWeight = FontWeight.W400,
-                fontSize = 12.sp
+                color = LocalPallet.current.text12,
+                style = LocalTypography.current.subtitleR12
             )
         }
         if (synchronizationContent != null) {
@@ -95,9 +90,8 @@ private fun ColumnScope.ComposableKeyCardContent(
         text = flipperKeyParsed.keyName,
         overflow = TextOverflow.Ellipsis,
         maxLines = 1,
-        color = colorResource(DesignSystem.color.black_100),
-        fontSize = 14.sp,
-        fontWeight = FontWeight.W400
+        color = LocalPallet.current.text100,
+        style = LocalTypography.current.bodyR14
     )
 
     val notes = flipperKeyParsed.notes ?: return
@@ -110,9 +104,8 @@ private fun ColumnScope.ComposableKeyCardContent(
                 end = 8.dp
             ),
             text = notes,
-            fontSize = 14.sp,
-            fontWeight = FontWeight.W400,
-            color = colorResource(DesignSystem.color.black_30),
+            style = LocalTypography.current.bodyR14,
+            color = LocalPallet.current.text30,
             maxLines = 1,
             overflow = TextOverflow.Ellipsis
         )